moved StylableBadge to deprecated folder, and created fake StylableBadge pointer for backward compatibility
