+++ packages/wix-ui-core/src/components/StylableBadge/index.ts
@@ -1,1 +1,2 @@
-export {Badge, BadgeProps} from './Badge';
+// Backward compatibility for deprecated components
+export * from '../../components/deprecated/StylableBadge';
