+++ packages/wix-ui-core/src/components/deprecated/StylableBadge/Badge.protractor.driver.ts
@@ -1,4 +1,4 @@
-import {BaseDriver, DriverFactory} from './../../common/BaseDriver.protractor';
+import {BaseDriver, DriverFactory} from './../../../common/BaseDriver.protractor';
 
 export interface BadgeDriver extends BaseDriver {
   text: () => Promise<string>;
