+++ packages/wix-ui-core/src/components/deprecated/StylableBadge/BadgeStyle.st.css
