+++ packages/wix-ui-core/src/components/deprecated/StylableBadge/index.ts
@@ -1,0 +1,1 @@
+export {Badge, BadgeProps} from './Badge';
