+++ packages/wix-ui-core/src/testkit/index.ts
@@ -74,7 +74,7 @@ export { CircularProgressBarDriver };
 import {
   badgeDriverFactory as stylableBadgeDriverFactory,
   BadgeDriver as StylableBadgeDriver
-} from "../components/StylableBadge/Badge.driver";
+} from "../components/deprecated/StylableBadge/Badge.driver";
 export const stylableBadgeTestkitFactory = testkitFactoryCreator<
   StylableBadgeDriver
 >(stylableBadgeDriverFactory);
