+++ packages/wix-ui-core/stories/StylableBadge.story.tsx
@@ -1,10 +1,10 @@
-import {Badge} from '../src/components/StylableBadge';
+import {Badge} from '../src/components/deprecated/StylableBadge';
 
 export default {
   category: 'Components',
   storyName: 'StylableBadge',
   component: Badge,
-  componentPath: '../src/components/StylableBadge/Badge.tsx',
+  componentPath: '../src/components/deprecated/StylableBadge/Badge.tsx',
 
   componentProps: {
     children: 'I\'m a Badge!',
